add modification, delete confirmation for food
